requirements.txt
Pillow
discord.py
python-dotenv
jishaku
requests